chore: Add Firebase Hosting deployment workflow
This commit adds a new workflow file, `firebase-hosting-merge.yml`, which is responsible for deploying the project to Firebase Hosting whenever there is a push event. The workflow runs on the latest version of Ubuntu and includes steps to checkout the code, install dependencies, build the project, and deploy it using the Firebase Hosting Deploy action. The deployment is configured to use the live channel and the project ID `tfg-nico-ruben`.