feat: configurar GitHub actions (#24)
* feat: conected login to backend and adapting frontend to the new react sintax

* feat: Added authentication context and protected routes

* fix: solving codacy bad issues

* fix: solving web visuality error

* feat: connected register to backend

Co-authored-by: Nicolás Sánchez Mendoza <nicsanmen@users.noreply.github.com>

* refactor: refactoring existen register component

* chore: Implicit converting the user input birth date to match the server's format

* feat: Updated RegisterPage rol select styling

The commit message suggests that the styling of the rol select element in the RegisterPage component has been updated. The previous className has been replaced with a new one that includes flexbox properties, border, padding, and width settings. The changes aim to improve the visual appearance and responsiveness of the select element.

* refactor: Remove RegisterPage component

The RegisterPage component has been removed from the project. This component was no longer needed and has been deleted to simplify the codebase.

* chore: added integration test with cypress cloud and coverage to codacy + gactions

* refactor: Remove unused matrix configuration in tests.yml

The `matrix` configuration in the `tests.yml` file was removed as it was no longer needed. This change simplifies the codebase and improves readability.

* chore: Add Firebase Hosting deployment workflow

This commit adds a new workflow file, `firebase-hosting-merge.yml`, which is responsible for deploying the project to Firebase Hosting whenever there is a push event. The workflow runs on the latest version of Ubuntu and includes steps to checkout the code, install dependencies, build the project, and deploy it using the Firebase Hosting Deploy action. The deployment is configured to use the live channel and the project ID `tfg-nico-ruben`.

---------

Co-authored-by: Nicolás Sánchez Mendoza <nicsanmen@users.noreply.github.com>