feat: configure google analytics (#35)
* fix: solving e2e tests error and trying to solve codacy bug

* feat: Add private route for dashboard page

The code changes include adding a private route for the dashboard page, which ensures that only authenticated users can access it. This improves the security and privacy of the dashboard functionality.

Co-authored-by: Nicolás Sánchez Mendoza <nicsanmen@users.noreply.github.com>

* feat: Update SideBar component with responsive design

The code changes in SideBar.jsx improve the responsiveness of the SideBar component. The component now adapts to different screen sizes, with a collapsible menu for smaller screens and a fixed layout for larger screens. This enhances the user experience and ensures that the SideBar remains usable on various devices.

Co-authored-by: Nicolás Sánchez Mendoza <nicsanmen@users.noreply.github.com>

* feat: added create event when new owner is register and visit the restaurant url

Co-authored-by: Nicolás Sánchez Mendoza <nicsanmen@users.noreply.github.com>

* feat: Set firebase environment variables for GitHub Actions workflow and update API URL

The code changes in `.github/workflows/tests.yml` add a step to set environment variables for the GitHub Actions workflow. This ensures that the necessary API keys and credentials are available during the workflow execution.

In addition, the code changes in `src/config.js` update the API URL to use the production server instead of the local development server. This ensures that the application connects to the correct backend API.

* fix: solving redireccion bug when token expired

* feat: added some restaurants analytics and a barchar for displaying the analitics

Co-authored-by: Nicolás Sánchez Mendoza <nicsanmen@users.noreply.github.com>

* fix: solving error navigator

* feat: Add LineChar component for displaying visits by range

The code changes in `Dashboard.jsx` add the `LineChar` component to display the visits by range on the dashboard page. This component enhances the analytics functionality by providing a visual representation of the visits over a specified range of days.

* feat: Update LineChar component for responsive design

The code changes in LineChar.jsx update the LineChar component to make it responsive. The width and height of the chart are now dynamically set based on the screen size. This ensures that the chart displays properly on both desktop and mobile devices.

Co-authored-by: Nicolás Sánchez Mendoza <nicsanmen@users.noreply.github.com>

* feat: Update API URL to use production server

The code changes in `src/config.js` update the API URL to use the production server instead of the local development server. This ensures that the application connects to the correct backend API.

---------

Co-authored-by: Nicolás Sánchez Mendoza <nicsanmen@users.noreply.github.com>