feat: Add support for displaying allergen images in menu
The code add support for displaying allergen images in the menu. The `Menu` component now renders the `AllergenImage` component for each allergen associated with a dish. The `Modal` component includes the `allergens` field and its corresponding value in the form. This enhancement improves the user experience by providing visual cues for allergens present in the dishes.